+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * templates: Add "immutable" property for parameters that should not be edited on application updates
 * Allow sending a list of applications to delete when sending a DELETE to /applications
 * driver: Add timestamps to tasks
+* Return applications using a template when querying the /templatesets endpoint
 
 ## Fixed
 * GUI: Fix task order in the Deploy Log when using AS3 versions 3.26.0 and 3.27.0